\name{facepp-package}
\alias{facepp-package}
\alias{facepp}
\docType{package}
\title{
\packageTitle{facepp}
}
\description{
\packageDescription{facepp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{facepp}
\packageIndices{facepp}
With a Face++ API key, you can use this package to access Face++ platform without pain.
~~
Before using the API, create a proxy object with \code{newFaceppProxy}. This object will store the 
tedious API key and secrets for you. 
}

\author{
\packageAuthor{facepp}

Maintainer: \packageMaintainer{facepp}
}

