\name{newFaceppProxy}
\alias{newFaceppProxy}
\title{
  newFaceppProxy
}
\description{
Create a Face++ API access proxy which stores API key and API secret.
}
\usage{
newFaceppProxy(api_key, api_secret)
}
\arguments{
  \item{api_key}{
    API key provided by Face++ platform.
  }
  \item{api_secret}{
   API secret provided by Face++ platform.
}
}
\value{
 A FaceppProxy object.
}

\author{
  Mengxiao Lin <linmx0130@gmail.com>
}
